# [mutators] deepcopy cached value instead of unmarshalling json