Instead of keeping the assigned value for AssignMutator/AssignMetadataMutator as json, cache the unmarshalled value.

Unmarshalling JSON is incredibly expensive, even for simple values. We should avoid this when possible. Instead, we should cache the value to assign, and then deepcopy it when needed. Fortunately, the runtime package provides us a method just for this!

For AssignMetadata we just cache the string value since it is only valid to assign strings with an AssignMetadataMutator.

This change results in a ~90% reduction in compute time for most common uses. The existing benchmarks are for assigning a string, but the speedup is even more for complex assigned values.

Change in AssingMutator benchmarks:
```
$ benchstat before.txt after.txt 
name                                            old time/op  new time/op  delta
$ benchstat before.out after.out 
name                                            old time/op  new time/op  delta
AssignMutator_Mutate/always_mutate_1-depth-12   2.93µs ± 4%  0.15µs ± 6%  -94.73%  (p=0.000 n=20+20)
AssignMutator_Mutate/always_mutate_2-depth-12   3.04µs ± 4%  0.18µs ± 7%  -93.99%  (p=0.000 n=20+20)
AssignMutator_Mutate/always_mutate_5-depth-12   3.28µs ± 3%  0.26µs ± 5%  -91.94%  (p=0.000 n=19+20)
AssignMutator_Mutate/always_mutate_10-depth-12  3.66µs ± 3%  0.40µs ± 6%  -89.10%  (p=0.000 n=19+20)
AssignMutator_Mutate/always_mutate_20-depth-12  4.31µs ± 3%  0.64µs ± 5%  -85.07%  (p=0.000 n=20+20)
AssignMutator_Mutate/never_mutate_1-depth-12     117ns ± 3%   116ns ± 5%     ~     (p=0.744 n=19+20)
AssignMutator_Mutate/never_mutate_2-depth-12     154ns ± 6%   153ns ± 5%     ~     (p=0.607 n=17+18)
AssignMutator_Mutate/never_mutate_5-depth-12     225ns ± 5%   228ns ± 6%     ~     (p=0.175 n=17+20)
AssignMutator_Mutate/never_mutate_10-depth-12    323ns ± 6%   329ns ± 4%     ~     (p=0.185 n=20+20)
AssignMutator_Mutate/never_mutate_20-depth-12    505ns ± 7%   510ns ± 5%     ~     (p=0.354 n=19+20)
AssignMetadataMutator_Always-12                 4.76µs ± 1%  1.93µs ± 1%  -59.54%  (p=0.000 n=19+19)
AssignMetadataMutator_Never-12                  54.2ns ± 0%  54.3ns ± 2%     ~     (p=0.859 n=15+17)
```

Signed-off-by: Will Beason <willbeason@google.com>