@willbeason 2021-07-19
Made the change for AssignMetadataMutator as well.